[FIX] : fixed axios error in MyGameLog
